Structure creation cleanup
