+++ src/training/input_recording/mod.rs
@@ -1,0 +1,3 @@
+//use smash::app::{self};
+
+pub mod structures;
